Bugfix/sticky-header-avatar-skipping
Hi
*Symptoms*:
The height of the #maintop container gets changed for the .sticktotop action when scrolling down.
When scrolling up, the transition makes the avatar picture to skip from left to right very fast it breaks the app experience. 
I guess the skipping is due he logo height change. 
*Fix*:
remove transition from the maintop container. This fixes it, it changes the header behaviour a little bit, but looks much more acceptable. 

*Possible refactor*
the logo can be wrapped inside some container with defined height and `overflow-y: visible`, etc.